// src/Components/types/FilterType.ts
export enum FilterType {
  AllPhones = 'allphones',
  Allphabetically = 'allphabetically',
  AllphabeticallyDescending = 'allphabeticallydescending',
  TheNewest = 'newest',
  TheOldest = 'oldest',
  TheMostExpensive = 'mostexpensive',
  TheCheapest = 'thecheapest',
  GoldColor = 'goldcolor',
  GraphiteColor = 'graphitecolor',
  SierrablueColor = 'sierrabluecolor',
  BlackColor = 'blackcolor',
  SpaceBlackColor = 'spaceblackcolor',
  RedGoldColor = 'redgoldcolor',
  RedColor = 'red',
  WhiteColor = 'white',
  YellowColor = 'yellow',
  Capacity_32_GB = 'capacity32gb',
  Capacity_64_GB = 'capacity64gb',
  Capacity_128_GB = 'capacity128gb',
  Capacity_256_GB = 'capacity256gb',
  Capacity_512_GB = 'capacity512gb',
  Capacity_1_TB = 'capacity1tb',
}